requirements.txt
flask==2.3.3
kafka-python==2.0.2
sqlalchemy==2.0.20
pg8000==1.30.2
psycopg2-binary==2.9.7
uuid==1.30
pytest==7.4.0
Werkzeug==2.3.7
jaydebeapi==1.2.3
jpype1==1.4.1
psycopg[binary]==3.1.12 